css/css-anchor-position/anchor-position-auto-002.html

<!DOCTYPE html>
<title>Tests evaluation of invalid auto anchor positioning</title>
<link rel="help" href="https://drafts4.csswg.org/css-anchor-1/#anchor-auto">
<link rel="auto" href="mailto:xiaochengh@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>

<style>
body {
  margin: 0;
}

#cb {
  position: absolute;
  width: 500px;
  height: 500px;
}

#anchor {
  margin-left: 150px;
  margin-top: 250px;
  width: 100px;
  height: 100px;
  background: orange;
  anchor-name: --a;
}

.target {
  position: absolute;
  width: 100px;
  height: 100px;
  background: lime;
  inset: 0; /* makes auto anchor position invalid */
}

#target1 {
  top: anchor(--a auto, 400px);
  left: anchor(--a auto-same, 200px);
}

#target2 {
  width: auto;
  height: auto;
  bottom: anchor(--a auto, 400px);
  right: anchor(--a auto-same, 300px);
}

#target3 {
  top: anchor(--a auto-same, 200px);
  left: anchor(--a auto, 300px);
}

#target4 {
  width: auto;
  height: auto;
  bottom: anchor(--a auto-same, 300px);
  right: anchor(--a auto, 400px);
}
</style>

<body onload="checkLayoutForAnchorPos('.target')">
  <div id="cb">
    <div id="anchor"></div>

    <div id="target1" class="target"
         data-offset-x="200" data-offset-y="400"></div>
    <div id="target2" class="target"
         data-expected-width="200" data-expected-height="100"></div>
    <div id="target3" class="target"
         data-offset-x="300" data-offset-y="200"></div>
    <div id="target4" class="target"
         data-expected-width="100" data-expected-height="200"></div>
  </div>
</body>
